+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,10 +1,11 @@
-# v.0.10.0 (2022-03-13)
+# v.0.17.0 (2022-03-13)
 
 - upgraded to "@cosmos-client/core": "0.44.4"
+- xchain-cosmos and xchain-thorchain now extend BaseXChainClient
 
 ## Breaking changes
 
-- Remove ``
+- Remove `minheight` and `maxheight` params from `CosmosSDKClient.searchTx` (params were removed from the API)
 
 # v.0.16.1 (2022-02-04)